+++ modules/graphics/functions/colormaps/cool.m
@@ -1,0 +1,22 @@
+%=============================================================================
+% Copyright (c) 2016-present Allan CORNET (Nelson)
+%=============================================================================
+% This file is part of the Nelson.
+%=============================================================================
+% LICENCE_BLOCK_BEGIN
+% SPDX-License-Identifier: LGPL-3.0-or-later
+% LICENCE_BLOCK_END
+%=============================================================================
+function map = cool(m)
+  if nargin < 1
+    f = get(groot(), 'CurrentFigure');
+    if isempty(f)
+      m = 256;
+    else
+      m = size(f.Colormap,1);
+    end
+  end
+  r = (0:m-1)' / max(m - 1, 1);
+  map = [r, (1 - r), ones(m, 1)]; 
+end
+%=============================================================================